Create motor speed control according to Temperature
here the NTC measure the heat in the inverter if the heat exceed a certain temperature the motor speed will be reduced gradually till it stop totally. it is in the Electrical Vehicles Systems, the Vehicle will turn to safety mode and reduce its speed gradually by controlling the inverter.